jalur_pendaftaran-scraper_2/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
openpyxl==3.1.5
tldextract==5.1.2
google-genai==1.3.0
playwright==1.49.0
pypdf==4.3.1